backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.2
supabase==2.3.4
langchain==0.1.0
langchain-community==0.0.13
langchain-huggingface==0.0.5
